♻️ Refactor folder structure
