--- client/src/layout/Library/components/LibraryHeader.jsx
@@ -1,7 +1,0 @@
-import styles from './LibraryHeader.module.scss';
-
-const LibraryHeader = () => {
-  return <div className={styles.header}>Header</div>;
-};
-
-export default LibraryHeader;
